Update README.md
Added more recent notes of much importance. It might not be abundantly clear how these notes on my many concepts across my novelistic phenomenologies is relevant to the question of "Cultural Software". It's really just a question of making indelible links between my Many Repositories, for future reference, as documentation of all my processes and practices, including my the "Complex Conceptual Systems Underlying My Many Interdisciplinary Art-Research Projects", which as of now are all subsumed under, and "Subprojects" of, the largest-scale Work of My Life, "THE PAINTER'S ALMANACH". More on this soon...