Update README.md
I removed a bunch of more creative writing that I had originally included in the README.md Markdown file in this GitHub repository, which I now find irrelevant to the conception and development of the concept of "Cultural Software", and only served to detract from it. I have saved the text in other places, for future reference, since it contained many great ideas from my more creative work.